+++ oddEvenMode.cc
@@ -1,7 +1,7 @@
 #include "oddEvenMode.h"
 
-std::vector<unsigned int> OddEvenMode::getOrder(std::vector<std::shared_ptr<Playable>> list) {
-    std::vector<unsigned int> order;
+std::vector<size_t> OddEvenMode::getOrder(std::vector<std::shared_ptr<Playable>> list) {
+    std::vector<size_t> order;
     for (size_t odd = 1; odd < list.size(); odd += 2) {
 			order.push_back(odd);
     }
