+++ oddEvenMode.h
@@ -6,7 +6,7 @@
 
 class OddEvenMode : public Mode {
 public:
-    std::vector<unsigned int> getOrder(std::vector<std::shared_ptr<Playable>> list) override;
+    std::vector<size_t> getOrder(std::vector<std::shared_ptr<Playable>> list) override;
     OddEvenMode() = default;
 };
