+++ shuffleMode.h
@@ -12,7 +12,7 @@ private:
     int seed;
 public:
     explicit ShuffleMode(int seed) : engine(std::default_random_engine(seed)), seed(seed) {};
-    std::vector<unsigned int> getOrder(std::vector<std::shared_ptr<Playable>> list) override;
+    std::vector<size_t> getOrder(std::vector<std::shared_ptr<Playable>> list) override;
 };
